requirements.txt
Flask>=2.0
Flask-WTF>=1.0
WTForms>=3.0
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
python-dotenv>=0.21
email-validator>=1.3
Flask-Mail>=0.9.1
Flask-Bootstrap>=3.3.7.1
